% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savePlot.R
\name{changeLegend}
\alias{changeLegend}
\title{Edit plot legend}
\usage{
changeLegend(
  plot,
  toggleTitle = F,
  cycleColors = F,
  changeTitle = F,
  changeLabels = F
)
}
\arguments{
\item{plot}{A ggplot object}

\item{toggleTitle}{a}

\item{cycleColors}{a}

\item{changeTitle}{a}

\item{changeLabels}{a}
}
\value{
Plot with edited legend
}
\description{
Edit plot legend
}
