% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeModel.R
\name{changeModel}
\alias{changeModel}
\title{Add or Remove predictors for a model}
\usage{
changeModel(model, remove = c(), add = c())
}
\arguments{
\item{model}{Model of class "aov", "lm", "glm", or "mlm".}

\item{remove}{Vector of predictor(s) to remove from the model.}

\item{add}{Vector of predictor(s) to add to the model.}
}
\value{
Model with specified predictors removed and added.
}
\description{
Add or Remove predictors for a model
}
