% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddata.R
\name{cleanData}
\alias{cleanData}
\title{Clean data by re-classing columns}
\usage{
cleanData(data)
}
\arguments{
\item{data}{}
}
\value{
Data.frame with columns classed as either "factor", "numeric", or "Date"
}
\description{
Clean data by re-classing columns
}
