% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddata.R
\name{cleanFactors}
\alias{cleanFactors}
\title{Clean up the factor variables in the data}
\usage{
cleanFactors(data, variables = NULL, exclude = NULL, all = F)
}
\arguments{
\item{data}{Data}

\item{variables}{(Optional) Names of variables to rename.
If not specified, will select all categorical variables in the data.}

\item{exclude}{(Optional) Names of variables to exclude.
If this and "variables" are not specified, will select all categorical variables in the data.}

\item{all}{Whether or not to refactor all variables without asking before each one. Defaults to False.}
}
\value{
Dataframe with cleaned up factor variables
}
\description{
Clean up the factor variables in the data
}
