% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{descriptive_stats}
\alias{descriptive_stats}
\title{Calculate statistics for descriptive characteristics between groups}
\usage{
descriptive_stats(
  data,
  characteristic,
  group_by,
  var_type = c("categorical", "continuous"),
  paired = F
)
}
\arguments{
\item{data}{Dataframe of data.}

\item{characteristic}{Column name in data that specifies the variable being tested between groups.}

\item{group_by}{Column name in data that specifies a factor to group data. Usually the independent variable of interest.}

\item{var_type}{Type of variable specified by "characteristic". Must be either "categorical" or "continuous".}

\item{paired}{Whether or not samples in each group are paired or unpaired. Defaults to False.}
}
\value{
Statistics summary using Chi-squared test for categorical variables
   or the appropriate statistical test for continuous variables
}
\description{
Calculate statistics for descriptive characteristics between groups
}
