% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savePlot.R
\name{editAxes}
\alias{editAxes}
\title{Edit plot axes}
\usage{
editAxes(
  plot,
  x.toggleTitle = F,
  x.toggleLabels = F,
  x.changeTitle = F,
  x.changeLabels = F,
  y.toggleTitle = F,
  y.toggleLabels = F,
  y.changeTitle = F,
  y.changeLabels = F
)
}
\arguments{
\item{plot}{A ggplot object}

\item{x.toggleTitle}{Toggle whether or not the x-axis title appears}

\item{x.toggleLabels}{Toggle whether or not the x-axis labels appear}

\item{x.changeTitle}{Change the x-axis title and font size}

\item{x.changeLabels}{Change the x-axis labels and font size}

\item{y.toggleTitle}{Toggle whether or not the y-axis title appears}

\item{y.toggleLabels}{Toggle whether or not the y-axis labels appear}

\item{y.changeTitle}{Change the y-axis title and font size}

\item{y.changeLabels}{Change the y-axis labels and font size}
}
\value{
New plot with edited axes
}
\description{
Edit plot axes
}
