% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOptModel.R
\name{findOptLM}
\alias{findOptLM}
\title{Find an optimized linear model for an outcome variable from a list of predictors}
\usage{
findOptLM(
  data = NA,
  outcome = c(),
  predictors = c(),
  family = "binomial",
  variable_of_interest = c(),
  include = c(),
  exclude = c(),
  dim_ratio = 10,
  dim_ratio_lax = 1
)
}
\arguments{
\item{data}{Data frame with patient IDs as rows and covariates as columns.}

\item{outcome}{Outcome variable. Should be the column name of the desired outcome variable in @data.}

\item{predictors}{Vector of predictor variables to select from when building the model. These should be column names in @data.}

\item{family}{Used in the glm function: "Type of error distribution and link function to be used in the model. For glm this can be a character string naming a family function, a family function or the result of a call to a family function. For glm.fit only the third option is supported. (See family for details of family functions.)". Defaults to binomial.}

\item{variable_of_interest}{(Optional) Variable of interest that from @predictors that must be included in the model.}

\item{include}{(Optional) Vector of predictor variable(s) listed in @predictors that the function should include first (after @variable_of_interest) when building the model.}

\item{exclude}{(Optional) Vector of predictor variable(s) listed in @predictors to exclude using when building the model.}

\item{dim_ratio}{Number used in the pruning step to determine the highest allowed dimensionality of the model (how many covariates the model is allowed to have on the right-hand side). The number of events in the data divided by @dim_ratio (rounded down to the nearest whole number) yields the highest dimensionality allowed. Defaults to 10.}

\item{dim_ratio_lax}{How many more dimensions is the model allowed to have than that calculated using @dim_ratio_lax. For example, for a dataset with 26 events and a @dim_ratio of 10, 26/10 = 2.6. A @dim_ratio_lax of 0 in this case would prune a model to 2 dimensions, while a @dim_ratio_lax of 1 would prune the model to 3 dimensions (assuming a model with at least 3 dimensions was able to be built).}
}
\value{
Returns a model from the @data with the specified @outcome variable and @variable_of_interest, trying to include as many of the @predictors specified in @include, and excluding any @predictors specified in @exclude, pruned using @dim_ratio and @dim_ratio_lax as explained above.
}
\description{
Find an optimized linear model for an outcome variable from a list of predictors
}
