% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOptModel.R
\name{findOptModel}
\alias{findOptModel}
\title{Title}
\usage{
findOptModel(
  data = NA,
  outcome = c(),
  predictors = c(),
  family = "binomial",
  variable_of_interest = c(),
  include = c(),
  exclude = c(),
  dim_ratio = 10,
  dim_ratio_lax = 1
)
}
\arguments{
\item{dim_ratio_lax}{}
}
\description{
Title
}
