% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFreq.R
\name{getFillColors}
\alias{getFillColors}
\title{Get "fill" colors from a ggplot object}
\usage{
getFillColors(plot)
}
\arguments{
\item{plot}{A ggplot object}
}
\value{
Returns the colors used in the "fill" of the ggplot
}
\description{
Get "fill" colors from a ggplot object
}
