% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHRs.R
\name{getHRs.base}
\alias{getHRs.base}
\title{Base Function for Hazard Ratio Calculation}
\usage{
getHRs.base(data, event.time, event.status, predictor_formula, repeatVar = F)
}
\arguments{
\item{data}{Data table in data.frame format}

\item{event.time}{Name, in string format, of column in data containing time-to-event values}

\item{event.status}{Name, in string format, of column in data containing event status values (as either 0/1 or T/F)}

\item{predictor_formula}{Arithmetic combination of desired predictor variables in string format. Must use column names in data}

\item{repeatVar}{Whether to repeat the variable name on the left-most column next to each category for categorical variables. Defaults to "FALSE"}
}
\value{
Returns a dataframe of Hazard Ratios and associated confidence intervals and p-values
}
\description{
Base Function for Hazard Ratio Calculation
}
