% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{getModelVars}
\alias{getModelVars}
\title{Function for obtaining list of variables from a model}
\usage{
getModelVars(model)
}
\arguments{
\item{model}{A model of any class containing a "model" attribute}
}
\value{
Returns a named list of the response variable (Response) and the independent variables (Terms).
}
\description{
Function for obtaining list of variables from a model
}
