% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getORs.R
\name{getOR.manual}
\alias{getOR.manual}
\title{Get the Odds Ratio for a single variable directly from the data}
\usage{
getOR.manual(data, variable, outcome, alpha = 0.05)
}
\arguments{
\item{data}{Data table in data.frame format}

\item{variable}{Name of column, in string format, for the variable of interest. Must also be 0/1 or T/F values}

\item{outcome}{Name of column, in string format, containing outcome values (as either 0/1 or T/F)}

\item{alpha}{Significance level. Defaults to 0.05}
}
\value{
Dataframe of Odds Ratio and associated confidence interval and p-value
}
\description{
Get the Odds Ratio for a single variable directly from the data
}
