% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getORs.R
\name{getORs.LM}
\alias{getORs.LM}
\title{Get Odds Ratios with confidence intervals and p-values from a linear model}
\usage{
getORs.LM(model, longer = F, repeatVar = F)
}
\arguments{
\item{model}{Model of class "aov", "lm", "glm", or "mlm".}

\item{longer}{Whether to format the table into a longer format or nested format. Defaults to a nested format ("FALSE").}

\item{repeatVar}{Whether to repeat the variable name on the left-most column next to each category for categorical variables. Defaults to "FALSE".}
}
\value{
Dataframe of odds ratios for each predictor in the model.
}
\description{
Get Odds Ratios with confidence intervals and p-values from a linear model
}
