% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getORs.R
\name{getORs}
\alias{getORs}
\title{Get Odds Ratios with confidence intervals and p-values from a linear model}
\usage{
getORs(model, repeatVar = F, longer = F)
}
\arguments{
\item{repeatVar}{}
}
\description{
Get Odds Ratios with confidence intervals and p-values from a linear model
}
