% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddata.R
\name{loadData}
\alias{loadData}
\title{Load Data from Excel or CSV File}
\usage{
loadData(path, verbose = T)
}
\arguments{
\item{path}{Path to f}

\item{verbose}{}
}
\value{
Data.frame of data loaded from Excel or CSV file.
}
\description{
Load Data from Excel or CSV File
}
