% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeModel.R
\name{makeModel}
\alias{makeModel}
\title{Make a Model}
\usage{
makeModel(data, response, time = NULL, predictors = c(), family = "binomial")
}
\arguments{
\item{data}{Data frame with patient IDs as rows and covariates as columns.}

\item{response}{Response variable. Should be the column name of the desired response variable in @data.}

\item{time}{For survival data, the time variable corresponding to the response.}

\item{predictors}{Vector of predictor/independent variables to select from when building the model. These should be column names in @data.}

\item{family}{Used in the glm function: "Type of error distribution and link function to be used in the model. For glm this can be a character string naming a family function, a family function or the result of a call to a family function. For glm.fit only the third option is supported. (See family for details of family functions.)". Defaults to binomial.}
}
\value{
A model.
}
\description{
Make a Model
}
