% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotForestPlot.R
\name{plotHRs}
\alias{plotHRs}
\title{Plot HRs}
\usage{
plotHRs(
  data,
  event.time,
  event.status,
  predictor_formula,
  n_breaks = 7,
  var.rename = c(),
  title = NULL,
  var_label_position = "top"
)
}
\arguments{
\item{data}{Data table in data.frame format.}

\item{event.time}{Name, in string format, of column in data containing time-to-event values.}

\item{event.status}{Name, in string format, of column in data containing event status values (as either 0/1 or T/F).}

\item{predictor_formula}{Arithmetic combination of desired predictor variables in string format. Must be column names in data.}

\item{n_breaks}{Number of axis ticks to create on the x-axis of the plot. Passed onto internal plotForestPlot function.}

\item{var.rename}{Named vector to rename variables with more appropriate names for a visual, if any need to be changed. For example, a column called "responded_to_treatment" could be renamed by inputing "c('responded_to_treatment'='Responded'). Passed onto internal plotForestPlot function.}

\item{title}{Title of the generated figure. Passed onto internal plotForestPlot function.}

\item{var_label_position}{When the plot is split by variables, this specifies whether the variable labels for each section of the plot are on the "top", "bottom", "right", or "left". Passed onto internal plotForestPlot function.}
}
\value{
Returns a forest plot of the hazard ratios for the predictors in the @model.
}
\description{
Plot HRs
}
