% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKM.R
\name{plotKM}
\alias{plotKM}
\title{Plot Kaplan-Meier plot of Events}
\usage{
plotKM(
  data,
  event.time,
  event.status,
  variable,
  time.unit = c("Days", "Weeks", "Months", "Years"),
  convert.time = c("Days", "Weeks", "Months", "Years"),
  event.name,
  showCI = T,
  showPVal = F,
  showRiskTab = F,
  offerSave = T,
  width = 8,
  height = 4
)
}
\arguments{
\item{data}{Data table in data.frame format}

\item{event.time}{Name, in string format, of column in @data containing time-to-event values}

\item{event.status}{Name, in string format, of column in @data containing event status values (as either 1/0 or T/F)}

\item{variable}{Variable of interest to separate data by into different curves. Must be a column name in @data}

\item{time.unit}{Specify how the time units are displayed on KM graph. Defaults to days}

\item{convert.time}{(Optional) Specify an alternate time unit to display on the KM plot}

\item{event.name}{(Optional) Name to display on y-axis title of KM plot. Will display as "@event.name Probability". Defaults to @event.status}

\item{showCI}{Whether or not to show confidence interval ribbons on KM plot. Defaults to True}

\item{showPVal}{Whether or not to show p-value on KM plot. Defaults to False}

\item{showRiskTab}{Whether or not to show risk table below KM plot. Defaults to False}

\item{offerSave}{Whether to offer saving the plot to a file. Defaults to True}

\item{width}{Width of figure, in inches. Decreasing width/height increases relative size of text. Defaults to 8}

\item{height}{Height of figure, in inches. Decreasing width/height increases relative size of text. Defaults to 4}
}
\value{
Returns a Kaplan-Meier plot
}
\description{
Plot Kaplan-Meier plot of Events
}
