% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddata.R
\name{refactor}
\alias{refactor}
\title{Add, remove, or change levels of a factor}
\usage{
refactor(variable)
}
\arguments{
\item{variable}{A factor to change the levels of}
}
\value{
Factor with refactored levels
}
\description{
Add, remove, or change levels of a factor
}
