% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddata.R
\name{relabel}
\alias{relabel}
\title{Change the values (not just the labels) of levels of factor variables in data}
\usage{
relabel(variable)
}
\arguments{
\item{variable}{Variable to relabel}
}
\value{
Relabeled variable
}
\description{
Change the values (not just the labels) of levels of factor variables in data
}
