% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddata.R
\name{renameVars}
\alias{renameVars}
\title{Rename Variables}
\usage{
renameVars(variables, auto = F)
}
\arguments{
\item{variables}{Vector of variables to rename.}

\item{auto}{Whether to automatically rename variables by replacing invalid characters only. Defaults to False.}
}
\value{
Named list of new variable names with names being the old variable names
}
\description{
Rename Variables
}
