% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddata.R
\name{saveData}
\alias{saveData}
\title{Save data to a csv file}
\usage{
saveData(data, folder, filename)
}
\arguments{
\item{data}{Data in table, dataframe, or tibble format}

\item{folder}{Path of folder to save data file to}

\item{filename}{Name to give the data file}
}
\value{
Nothing
}
\description{
Save data to a csv file
}
