% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savePlot.R
\name{savePlot}
\alias{savePlot}
\title{Save plot}
\usage{
savePlot(plot, width = NULL, height = NULL, separateLegend = F)
}
\arguments{
\item{plot}{Plot created by either plotFreq or plotKM. Must be a "ggplot" or "ggsurvfit" object}

\item{width}{Width of figure, in inches. Decreasing width/height increases relative size of text. Defaults to 10 for ggplots and 8 for ggsurvfit}

\item{height}{Height of figure, in inches. Decreasing width/height increases relative size of text. Defaults 8 for ggplots and 4 for ggsurvfit}

\item{separateLegend}{Whether or not to save legend separately to a different file. Defaults to False}
}
\value{
Saves the plot in "png" format and in standardized dimensions and resolution. Returns the path of the saved figure
}
\description{
Save plot
}
