% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{saveTab}
\alias{saveTab}
\title{Save a table as a CSV file}
\usage{
saveTab(table)
}
\arguments{
\item{table}{}
}
\value{
Saves a table to a CSV file in desired location and prints that location
}
\description{
Save a table as a CSV file
}
