% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{table1}
\alias{table1}
\title{Generate a "Table1" Table}
\usage{
table1(
  data,
  characteristics = c(),
  group_by,
  props_within_group = F,
  formula = NULL,
  exclude = c(),
  signif = 3,
  stats = F
)
}
\arguments{
\item{data}{Can be either a dataframe or a model containing a "model" attribute.}

\item{characteristics}{List of characteristics to be summarized in the table. Should be column names in data.}

\item{group_by}{Column name in data that specifies a factor to group summarizing statistics.}

\item{props_within_group}{Whether to calculate proportions of categorical variables within each group or across all groups.
Defaults to across all groups.}

\item{formula}{(Optional) Formula specifying which variables in data to show and any stratification if desired.}

\item{exclude}{(Optional) Variables in data that are to be excluded, if any.}

\item{signif}{Number of significant digits to report for statistical values. Defaults to 3.}

\item{stats}{Whether to test for significant difference between groups.}
}
\value{
Data.frame and optional saved csv of a summary of sample characteristics for the provided data.
}
\description{
Generate a "Table1" Table
}
