% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeModel.R
\name{updateModel}
\alias{updateModel}
\title{Wrapper for update() Function}
\usage{
updateModel(model, formula, ..., evaluate = T)
}
\arguments{
\item{model}{"An existing fit from a model function such as lm, glm and many others."}

\item{formula}{"Changes to the formula – see update.formula for details."}

\item{...}{"Additional arguments to the call, or arguments with changed values. Use name = NULL to remove the argument name."}

\item{evaluate}{"If true evaluate the new call else return the call."}
}
\value{
Updated model as updated by the update() function
}
\description{
Wrapper for update() Function
}
