% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModel.R
\name{buildModel}
\alias{buildModel}
\title{Find an optimized linear regression model}
\usage{
buildModel(
  data = NA,
  outcome = c(),
  predictors = c(),
  family = "binomial",
  variable_of_interest = c(),
  include = c(),
  exclude = c(),
  dim_ratio = 10,
  dim_ratio_lax = 1,
  returnAll = F
)
}
\arguments{
\item{returnAll}{}
}
\description{
Find an optimized linear regression model
}
