% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeModel.R
\name{changeModel}
\alias{changeModel}
\title{User-friendly Wrapper Function to Add or Drop Predictors for a Model}
\usage{
changeModel(model, drop = c(), add = c(), verbose = T)
}
\arguments{
\item{model}{Model of class "aov", "lm", "glm", "mlm", "coxph", and potentially other models. Passed to stats::update().}

\item{drop}{Vector of predictor(s) to drop from the model.}

\item{add}{Vector of predictor(s) to add to the model.}

\item{verbose}{Whether to print statements.}
}
\value{
Model with specified variables removed and added.
}
\description{
User-friendly Wrapper Function to Add or Drop Predictors for a Model
}
