% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModel.R
\name{find_best_predictor}
\alias{find_best_predictor}
\title{Find next best predictor for a model}
\usage{
find_best_predictor(
  data = NA,
  outcome = c(),
  predictors = c(),
  family = "binomial",
  baseline = c(),
  include = c(),
  exclude = c(),
  verbose = T
)
}
\arguments{
\item{verbose}{}
}
\value{
Outputs:
   improved (boolean)
   model.opt (optimized model)
   model.candidate (best model using the variables being tested)
   newpredictors.best (best variable)
   predictors.opt (set of all predictors used in model.opt)
   AICs (list of AICs for all predictors)
   LogLiks (list of LogLiks for all predictors)
}
\description{
Find next best predictor for a model
}
