% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHRs.R
\name{getHRs}
\alias{getHRs}
\title{Get Hazard Ratios}
\usage{
getHRs(
  data,
  event.time,
  event.status,
  predictors,
  subgroup_by = NULL,
  repeatVar = F,
  verbose = T
)
}
\arguments{
\item{data}{Data table in data.frame format}

\item{event.time}{Name, in string format, of column in data containing time-to-event values}

\item{event.status}{Name, in string format, of column in data containing event status values (as either 0/1 or T/F)}

\item{predictors}{Either a vector of 1 or more predictor variables or an arithmetic combination of predictor variables in character format}

\item{subgroup_by}{(Optional) Vector of variable names for subgroup analyses. Must be column names in data that are not in predictor_formula.}

\item{repeatVar}{Whether to repeat the variable name on the left-most column next to each category for categorical variables. Defaults to False}

\item{verbose}{Whether to print extra information during processing and analysis. Defaults to True.}
}
\value{
Returns a dataframe of Hazard Ratios and associated confidence intervals and p-values
}
\description{
Get Hazard Ratios
}
