% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getORs.R
\name{getORs}
\alias{getORs}
\title{Wrapper function for getting Odds Ratios}
\usage{
getORs(
  model = NULL,
  data = NULL,
  variable = NULL,
  response = NULL,
  alpha = 0.05,
  repeatVar = F,
  longer = F
)
}
\arguments{
\item{model}{Model of class "aov", "lm", "glm", or "mlm" to calculate OR from a linear model. Must either provide a model or @data with @variable and @response of interest}

\item{data}{Data table in data.frame format to calculate OR directly from the data using @variable and @response arguments}

\item{variable}{Name of column, in string format, for the variable of interest. Must also be 0/1 or T/F values}

\item{response}{Name of column, in string format, containing boolean response values (as either 0/1 or T/F)}

\item{alpha}{Significance level. Defaults to 0.05}

\item{repeatVar}{Whether to repeat the variable name on the left-most column next to each category for categorical variables. Defaults to "FALSE".}

\item{longer}{Whether to format the table into a longer format or nested format. Defaults to a nested format ("FALSE").}
}
\value{
Returns table of ORs
}
\description{
Wrapper function for getting Odds Ratios
}
