% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFreq.R
\name{plotFreq}
\alias{plotFreq}
\title{Plot response frequency (contingency table values) for a single binary variable}
\usage{
plotFreq(
  data,
  variable,
  response,
  variable_print = NULL,
  response_print = NULL,
  stratify_by = NULL,
  proportional = F,
  flip = F,
  separateLegend = F,
  includeNA = F,
  grp_colors = NA
)
}
\arguments{
\item{data}{Data table in data.frame format}

\item{variable}{Name of column, in string format, for the variable of interest. Must also be 0/1 or T/F values}

\item{response}{Name of column, in string format, containing boolean response values (as either 0/1 or T/F)}

\item{variable_print}{(Optional) Name of the variable to display on the figure if different from the column name in @data@}

\item{response_print}{(Optional) Name of the response to display on the figure if different from the column name in @data@}

\item{stratify_by}{(Optional) Name of column, in string format, for a third variable to stratify results and plot by}

\item{proportional}{Whether to display as absolute (if set to False) or proportional values (if set to True)}

\item{flip}{Whether to flip the x and y axes of the figure}

\item{separateLegend}{Whether or not to save legend separately to a different file. Defaults to False}

\item{includeNA}{Whether or not to include NAs in the figure}

\item{grp_colors}{(Optional) A list or named list of colors to use in the figure}
}
\value{
Plot in ggplot format
}
\description{
Plot response frequency (contingency table values) for a single binary variable
}
