% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotORs.R
\name{plotORs}
\alias{plotORs}
\title{Plot ORs}
\usage{
plotORs(
  model,
  n_breaks = 7,
  var.rename = c(),
  title = NULL,
  var_label_position = "right"
)
}
\arguments{
\item{title}{}
}
\description{
Plot ORs
}
