% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruneModel.R
\name{pruneModel}
\alias{pruneModel}
\title{Prune a model from a list of models provided}
\usage{
pruneModel(models, dim_ratio = 10, dim_ratio_lax = 1)
}
\arguments{
\item{dim_ratio_lax}{}
}
\description{
Prune a model from a list of models provided
}
