% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddata.R
\name{refactor}
\alias{refactor}
\title{Change the values of levels of factor variables in data}
\usage{
refactor(data, variables)
}
\arguments{
\item{data}{Data}

\item{variables}{Variables to refactor}
}
\value{
Dataframe with refactored variables
}
\description{
Change the values of levels of factor variables in data
}
