% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddata.R
\name{relabel}
\alias{relabel}
\title{Add or change labels of levels of a factor}
\usage{
relabel(factor)
}
\arguments{
\item{factor}{A factor to change the labels of}
}
\value{
Factor with relabeled levels
}
\description{
Add or change labels of levels of a factor
}
